refactor(stage): extract #410 reload guard into ndi_reload_guard module
ndi_watchdog.rs crossed the 1000-line hard cap after the #410 additions.
Move the pure decision functions (should_reload_given_pipeline_state,
healthz_body_has_streaming_pipeline), the /healthz fetch, and their unit
tests into a sibling ndi_reload_guard module (same split-for-size pattern the
file already documents). maybe_reload calls into it; no behavior change.
