+++ crates/presenter-ui/src/components/stage/ndi_watchdog.rs
@@ -253,7 +253,8 @@ impl ReloadEscalation {
         self.check_in_flight.set(true);
         let escalation = Rc::clone(self);
         spawn_local(async move {
-            let has_streaming = fetch_healthz_has_streaming_pipeline().await;
+            let has_streaming =
+                super::ndi_reload_guard::fetch_healthz_has_streaming_pipeline().await;
             escalation.check_in_flight.set(false);
             // A frame may have decoded while the check was in flight — re-check
             // the horizon so a recovered stream is never reloaded out from
@@ -266,7 +267,7 @@ impl ReloadEscalation {
             {
                 return;
             }
-            if !should_reload_given_pipeline_state(has_streaming) {
+            if !super::ndi_reload_guard::should_reload_given_pipeline_state(has_streaming) {
                 // Source legitimately down — reloading can't help. Reset the
                 // page-session timer so the rule re-evaluates after another
                 // full window instead of looping reloads every ~60s.
@@ -288,71 +289,6 @@ impl ReloadEscalation {
         });
         true
     }
-}
-
-/// LAST-RESORT reload gate (#410): given whether the SERVER currently has an
-/// actively-streaming NDI pipeline, should the stuck stage page reload?
-///
-/// - `true` (server is streaming, but this consumer decoded nothing) → reload:
-///   the page/consumer is stuck and a fresh WHEP negotiation + DOM can recover
-///   it (the #401 case).
-/// - `false` (no streaming pipeline) → DON'T reload: the source itself is down
-///   (Resolume silent / `ndi_pipelines` empty), so reloading cannot produce
-///   frames — it would just loop every ~60s. Keep waiting instead.
-///
-/// Pure + side-effect-free so the gate is unit-testable on host without a
-/// browser or a running server. The fetch-failure case is handled by the
-/// caller (it defaults to reloading), NOT here.
-pub(crate) fn should_reload_given_pipeline_state(server_has_streaming_pipeline: bool) -> bool {
-    server_has_streaming_pipeline
-}
-
-/// Parse a `/healthz` JSON body and decide whether the server has at least one
-/// actively-streaming NDI pipeline. An entry counts as "streaming" when its
-/// `state` is `"streaming"` or `"starting"` (about to deliver frames) — a
-/// `"stopped"` / `"errored"` pipeline, or an empty/absent `ndi_pipelines`
-/// array, means the source is NOT delivering media.
-///
-/// Pure over the response text so it is unit-testable on host. A body that
-/// fails to parse returns `false` here — but callers treat a FAILED fetch
-/// (network error, non-2xx) as "reload anyway", so a parse-of-garbage that
-/// only happens on a successful-but-malformed response stays conservative
-/// (no reload) rather than masking a real stuck consumer.
-pub(crate) fn healthz_body_has_streaming_pipeline(body: &str) -> bool {
-    let Ok(json) = serde_json::from_str::<serde_json::Value>(body) else {
-        return false;
-    };
-    let Some(pipelines) = json.get("ndi_pipelines").and_then(|v| v.as_array()) else {
-        return false;
-    };
-    pipelines.iter().any(|p| {
-        matches!(
-            p.get("state").and_then(|s| s.as_str()),
-            Some("streaming") | Some("starting")
-        )
-    })
-}
-
-/// Fetch `/healthz` and report whether the server has an actively-streaming NDI
-/// pipeline. Returns `true` on ANY fetch/response failure so a transient
-/// `/healthz` outage never SUPPRESSES a genuinely-needed last-resort reload
-/// (#410 — additive guard, must not break #401 recovery).
-async fn fetch_healthz_has_streaming_pipeline() -> bool {
-    async fn try_fetch() -> Option<bool> {
-        let window = leptos::web_sys::window()?;
-        let resp_val = JsFuture::from(window.fetch_with_str("/healthz"))
-            .await
-            .ok()?;
-        let resp: leptos::web_sys::Response = resp_val.dyn_into().ok()?;
-        if !resp.ok() {
-            return None;
-        }
-        let text = JsFuture::from(resp.text().ok()?).await.ok()?;
-        let body = text.as_string()?;
-        Some(healthz_body_has_streaming_pipeline(&body))
-    }
-    // Fetch/parse failure → default to TRUE (reload), preserving #401 behavior.
-    try_fetch().await.unwrap_or(true)
 }
 
 /// Watchdog that fires `on_failure` when EITHER:
@@ -1053,10 +989,7 @@ fn install_ice_failure_listener<F: Fn() + 'static>(
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        healthz_body_has_streaming_pipeline, should_escalate_reload,
-        should_reload_given_pipeline_state, Watchdog,
-    };
+    use super::{should_escalate_reload, Watchdog};
 
     // ─────────────────────────────────────────────────────────────────────
     // #401 LAST-RESORT page-reload escalation.
@@ -1125,72 +1058,5 @@ mod tests {
             Watchdog::RELOAD_NO_FRAME_MS,
             reconnect_path_budget,
         );
-    }
-
-    // ─────────────────────────────────────────────────────────────────────
-    // #410 server-liveness reload gate.
-    //
-    // The #401 last-resort reload assumed reloading could recover any stuck
-    // stream. But when the SOURCE itself is legitimately offline (Resolume
-    // silent → the server has NO streaming pipeline, /healthz ndi_pipelines is
-    // empty), no consumer can ever decode a frame, so the page reloaded every
-    // ~60s forever — benign but wasteful. The gate distinguishes
-    // "source legitimately down" (don't reload, keep waiting) from
-    // "my page/consumer is stuck while the server IS streaming" (reload, the
-    // #401 recovery). A failed /healthz fetch must NOT suppress a needed
-    // reload, so the caller defaults to reloading on fetch error.
-    // ─────────────────────────────────────────────────────────────────────
-
-    #[test]
-    fn reload_when_server_has_a_streaming_pipeline_but_consumer_is_stuck() {
-        // Server is streaming, but THIS page decoded nothing for the whole
-        // window → the consumer/page is stuck → reload (the #401 recovery).
-        assert!(should_reload_given_pipeline_state(true));
-    }
-
-    #[test]
-    fn suppress_reload_when_server_has_no_streaming_pipeline() {
-        // No streaming pipeline → the source itself is down → reloading cannot
-        // help → DON'T reload, keep waiting.
-        assert!(!should_reload_given_pipeline_state(false));
-    }
-
-    #[test]
-    fn healthz_with_streaming_pipeline_means_server_is_streaming() {
-        let body = r#"{"status":"ok","version":"0.4.138","channel":"dev",
-            "ndi_pipelines":[{"source_id":"abc","state":"streaming"}]}"#;
-        assert!(healthz_body_has_streaming_pipeline(body));
-    }
-
-    #[test]
-    fn healthz_with_starting_pipeline_counts_as_streaming() {
-        // A "starting" pipeline is about to deliver frames — treat it as live
-        // so we still attempt the consumer-stuck reload.
-        let body = r#"{"ndi_pipelines":[{"source_id":"abc","state":"starting"}]}"#;
-        assert!(healthz_body_has_streaming_pipeline(body));
-    }
-
-    #[test]
-    fn healthz_empty_pipelines_means_source_down() {
-        // The exact #410 symptom: source offline → empty ndi_pipelines.
-        let body = r#"{"status":"ok","ndi_pipelines":[]}"#;
-        assert!(!healthz_body_has_streaming_pipeline(body));
-    }
-
-    #[test]
-    fn healthz_only_errored_or_stopped_pipeline_means_source_down() {
-        let body = r#"{"ndi_pipelines":[
-            {"source_id":"a","state":"errored","last_error":"no source"},
-            {"source_id":"b","state":"stopped"}]}"#;
-        assert!(!healthz_body_has_streaming_pipeline(body));
-    }
-
-    #[test]
-    fn healthz_missing_field_or_garbage_means_not_streaming() {
-        // No ndi_pipelines key, or unparseable body → not streaming (the
-        // SUCCESSFUL-but-malformed case; a FAILED fetch is the caller's
-        // reload-anyway default, not this function's).
-        assert!(!healthz_body_has_streaming_pipeline(r#"{"status":"ok"}"#));
-        assert!(!healthz_body_has_streaming_pipeline("not json at all"));
     }
 }
